move test functions out of datastore interface
