+++ server/mock/datastore_mock.go
@@ -294,12 +294,6 @@ type CleanupExpiredHostsFunc func(ctx context.Context) ([]uint, error)
 
 type ScheduledQueryIDsByNameFunc func(ctx context.Context, batchSize int, packAndSchedQueryNames ...[2]string) ([]uint, error)
 
-type SaveQueryResultRowsFunc func(ctx context.Context, rows []*fleet.ScheduledQueryResultRow) error
-
-type QueryResultRowsFunc func(ctx context.Context, queryID uint, hostID uint) ([]*fleet.ScheduledQueryResultRow, error)
-
-type DeleteQueryResultsForHostFunc func(ctx context.Context, hostID uint, queryID uint) error
-
 type ResultCountForQueryFunc func(ctx context.Context, queryID uint) (int, error)
 
 type ResultCountForQueryAndHostFunc func(ctx context.Context, queryID uint, hostID uint) (int, error)
@@ -1108,15 +1102,6 @@ type DataStore struct {
 
 	ScheduledQueryIDsByNameFunc        ScheduledQueryIDsByNameFunc
 	ScheduledQueryIDsByNameFuncInvoked bool
-
-	SaveQueryResultRowsFunc        SaveQueryResultRowsFunc
-	SaveQueryResultRowsFuncInvoked bool
-
-	QueryResultRowsFunc        QueryResultRowsFunc
-	QueryResultRowsFuncInvoked bool
-
-	DeleteQueryResultsForHostFunc        DeleteQueryResultsForHostFunc
-	DeleteQueryResultsForHostFuncInvoked bool
 
 	ResultCountForQueryFunc        ResultCountForQueryFunc
 	ResultCountForQueryFuncInvoked bool
@@ -2676,27 +2661,6 @@ func (s *DataStore) ScheduledQueryIDsByName(ctx context.Context, batchSize int, 
 	s.ScheduledQueryIDsByNameFuncInvoked = true
 	s.mu.Unlock()
 	return s.ScheduledQueryIDsByNameFunc(ctx, batchSize, packAndSchedQueryNames...)
-}
-
-func (s *DataStore) SaveQueryResultRows(ctx context.Context, rows []*fleet.ScheduledQueryResultRow) error {
-	s.mu.Lock()
-	s.SaveQueryResultRowsFuncInvoked = true
-	s.mu.Unlock()
-	return s.SaveQueryResultRowsFunc(ctx, rows)
-}
-
-func (s *DataStore) QueryResultRows(ctx context.Context, queryID uint, hostID uint) ([]*fleet.ScheduledQueryResultRow, error) {
-	s.mu.Lock()
-	s.QueryResultRowsFuncInvoked = true
-	s.mu.Unlock()
-	return s.QueryResultRowsFunc(ctx, queryID, hostID)
-}
-
-func (s *DataStore) DeleteQueryResultsForHost(ctx context.Context, hostID uint, queryID uint) error {
-	s.mu.Lock()
-	s.DeleteQueryResultsForHostFuncInvoked = true
-	s.mu.Unlock()
-	return s.DeleteQueryResultsForHostFunc(ctx, hostID, queryID)
 }
 
 func (s *DataStore) ResultCountForQuery(ctx context.Context, queryID uint) (int, error) {
